Split the concern between using prebuilt binaries and uber JAR
Both capabilities cross-cutting and splitting them makes the code
simplier. Each capabilities can be used without the other one. In this
case, the getNativeRuntimeFiles() API was added to help simplify the
code. It will be released with 0.4. Basically, this new API is a
configurable file collection containing all the native runtime files to
be included into the JNI JAR. For single variant, the JNI and JVM JAR
are the same. It is possible to overwrite the collection via
`ConfigurableFileCollection.setFrom`.