Linux support (only Gnome for now)
This commit adds a Linux module to the plugin. Because Linux has a variety of desktop environments, this plugin uses a delegation strategy to allow a delegate ThemeMonitorService to be provided depending on the desktop environment being used. For now, there is only Gnome support.

It should be noted that this module accomplishes its task much differently than the Windows and Mac OS modules. Because Gnome (and others) support user themes, there is no centralized switch in the OS or its components for using "the" light theme or dark theme. It is common practice for user themes (at least for Gnome) to be provided in a light and dark form (along with a few others sometimes like "darker" and "darkest"). This is also explained in the JavaDoc for GtkVariants.

 Instead of using native OS API calls to ask the OS whether it is using the "dark theme" or the "light theme", the Gnome ThemeMonitorService deduces (based on the name of the current theme) whether the user's current theme is a light theme or a dark theme. This is a temporary measure I am using simply as a proof-of-concept that Gnome (and others in the future) can be supported by this plugin. If the PR for this feature gets the go-ahead, I intend to add an additional area to the look and feel settings which allows the user to choose which of their themes is their light, dark, and high contrast theme. This would replace the guessing functionality.