Fixed a bug with the default Gnome theme values
The settings menu wouldn't show the proper defaults. This was because the exact _instance_ of GtkTheme that the defaults held weren't part of the ChoiceProperty#choices. Because the default option wasn't valid among the choices, it would end up picking some random theme for the light, dark, and high contrast themes.